% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayHypOptParOnXGB_sepsisPatients.R
\name{BayHypOptParOnXGB_sepsisPatients}
\alias{BayHypOptParOnXGB_sepsisPatients}
\title{optimize hyperparameters}
\usage{
BayHypOptParOnXGB_sepsisPatients(snpdatafile, FUN, Kfold, bounds,
  init_points, n_iter, nrounds, early_stopping_rounds)
}
\description{
optimize hyperparameters
}
