% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayHypOptParOnXGBSNP.R
\name{BayHypOptParOnXGBSNP}
\alias{BayHypOptParOnXGBSNP}
\title{optimize hyperparameters}
\usage{
BayHypOptParOnXGBSNP(SNPdatafile, covariatesfile, CovariatesToInclude,
  Kfold, bounds, init_points, n_iter, nrounds, early_stopping_rounds)
}
\description{
optimize hyperparameters
}
