% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LightGBMOnSNPdata_cv.R
\name{LightGBMOnSNPdata_cv}
\alias{LightGBMOnSNPdata_cv}
\title{Use LightGBM to analyze SNP-data in a control/case GWAS setting using cross-validation in a HPC-environment
This function tries to find candidates of effects or interactions associated with a disease}
\usage{
LightGBMOnSNPdata_cv(TrainData, ValData, covariatesfile = NULL,
  CovariatesToInclude = NULL, Kround, hyperparameters, nrounds,
  early_stopping_rounds, nthreads, max_depth)
}
\arguments{
\item{covariatesfile}{is a table of non-genetic (PCs are OK) covariates for each IID.}

\item{CovariatesToInclude}{is a named vector of covariates (not including IID and SEX) to include.}

\item{hyperparameters}{is a list containing hyperparameters. See xgboost() package in R.}

\item{nrounds}{is the number of trees to be built during xgboost algorithm}

\item{early_stopping_rounds}{is the number of consecutive trees with no improvment in training error
before the algorithm stops to produce more trees.}

\item{Trainingdata}{is a vector of .raw-SNP-data files, with indivdual per row, and one column showing
the phenotype of interest and the rest of the columns are the genotype for spesific SNPs along the genome.}

\item{Kfold}{is the number of folds wanted in cross-validation}
}
\value{
The mean test error after cross validation of xgboost-algorithm on SNP-data.
}
\description{
Use LightGBM to analyze SNP-data in a control/case GWAS setting using cross-validation in a HPC-environment
This function tries to find candidates of effects or interactions associated with a disease
}
\examples{
LightGBMOnSNPdata_cv("~/directory/snpfile.txt",Kfold = 5, hyperparameters = list(eta = 0.3),
nrounds = 100, early_stopping_rounds = 10)
}
