% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoostOnSNPdata_cv.R
\name{XGBoostOnSNPdata_cv}
\alias{XGBoostOnSNPdata_cv}
\title{Use XGBoost to analyze SNP-data in a control/case GWAS setting using cross-validation in a HPC-environment
This function tries to find candidates of effects or interactions associated with a disease}
\usage{
XGBoostOnSNPdata_cv(TrainData, ValData, features = NULL,
  covariatesTable = NULL, Kround, hyperparameters, nrounds,
  early_stopping_rounds, nthreads)
}
\arguments{
\item{features}{is a file consisting of one column of features names to be used from TrainData and ValData}

\item{covariatesTable}{is a data frame or data table of non-genetic covariates for each IID.}

\item{hyperparameters}{is a list containing hyperparameters. See xgboost() package in R.}

\item{nrounds}{is the number of trees to be built during xgboost algorithm}

\item{early_stopping_rounds}{is the number of consecutive trees with no improvment in training error
before the algorithm stops to produce more trees.}

\item{Trainingdata}{is a vector of .raw-SNP-data files, with indivdual per row, and one column showing
the phenotype of interest and the rest of the columns are the genotype for spesific SNPs along the genome.}

\item{CovariatesToInclude}{is a named vector of covariates (not including IID and SEX) to include.}

\item{Kfold}{is the number of folds wanted in cross-validation}
}
\value{
The mean test error after cross validation of xgboost-algorithm on SNP-data.
}
\description{
Use XGBoost to analyze SNP-data in a control/case GWAS setting using cross-validation in a HPC-environment
This function tries to find candidates of effects or interactions associated with a disease
}
\examples{
XGBoostOnSNPdata_cv("~/directory/snpfile.txt",Kfold = 5, hyperparameters = list(eta = 0.3),
nrounds = 100, early_stopping_rounds = 10)
}
